requirements.txt
svg.path==4.1
Kivy==2.3.1
typing-extensions>=4.0.0
pytest>=7.0.0
black>=23.0.0